# docs/swagger.yaml
definitions:
  models.BasicListeningItem:
    properties:
      audio_files:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
      part_list:
        items:
          type: integer
        type: array
      status:
        type: integer
      type:
        type: integer
      user_id:
        type: string
    type: object
...
  models.ListeningItem:
    properties:
      audio_files:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
      part_list:
        items:
          $ref: '#/definitions/models.ListeningPartItem'
        type: array
      status:
        type: string
      type:
        type: string
    type: object
  models.ListeningListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ListeningItem'
        type: array
      total:
        type: integer
    type: object
  models.ListeningPartItem:
    properties:
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      type_list:
        items:
          $ref: '#/definitions/models.ListeningTypeItem'
        type: array
    type: object
  models.ListeningPartListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ListeningPartItem'
        type: array
      total:
        type: integer
    type: object
...
  models.ReadingPartItem:
    properties:
      article:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      type_list:
        items:
          $ref: '#/definitions/models.ReadingTypeItem'
        type: array
    type: object
  models.ReadingPartListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.ReadingPartItem'
        type: array
      total:
        type: integer
    type: object
...
  models.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /config/listening-part/add:
    post:
      consumes:
      - application/json
      description: 新增听力part
      parameters:
      - description: 听力part内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ListeningPartItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增听力part
      tags:
      - Listening
  /config/listening-part/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除听力part
      parameters:
      - description: 听力partID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除听力part
      tags:
      - Listening
  /config/listening-part/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取听力part详情
      parameters:
      - description: 听力partid
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ListeningPartItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取听力part详情
      tags:
      - Listening
  /config/listening-part/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取听力part列表，并返回分页结果
      parameters:
      - description: 试题名称
        in: query
        name: name
        type: string
      - description: 试题状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ListeningPartListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取听力篇列表
      tags:
      - Listening
  /config/listening-part/update:
    put:
      consumes:
      - application/json
      description: 更新听力part
      parameters:
      - description: 听力part内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ListeningPartItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新听力part
      tags:
      - Listening
  /config/listening/add:
    post:
      consumes:
      - application/json
      description: 新增听力套题
      parameters:
      - description: 听力套题内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.BasicListeningItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增听力套题
      tags:
      - Listening
  /config/listening/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除听力套题
      parameters:
      - description: 听力套题ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除听力套题
      tags:
      - Listening
  /config/listening/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取听力详情
      parameters:
      - description: 听力id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ListeningItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取听力套题详情
      tags:
      - Listening
  /config/listening/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取听力列表，并返回分页结果
      parameters:
      - description: 听力名称
        in: query
        name: name
        type: string
      - description: 听力状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ListeningListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取听力套题列表
      tags:
      - Listening
  /config/listening/update:
    put:
      consumes:
      - application/json
      description: 更新听力套题
      parameters:
      - description: 听力套题内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.BasicListeningItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新听力套题
      tags:
      - Listening
  /config/reading-part/add:
    post:
      consumes:
      - application/json
      description: 新增阅读part
      parameters:
      - description: 阅读part内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ReadingPartItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增阅读part
      tags:
      - Reading
  /config/reading-part/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除阅读part
      parameters:
      - description: 阅读partID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除阅读part
      tags:
      - Reading
  /config/reading-part/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取阅读part
      parameters:
      - description: 阅读part id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ReadingPartItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取阅读part详情
      tags:
      - Reading
  /config/reading-part/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取阅读part列表，并返回分页结果
      parameters:
      - description: 试题名称
        in: query
        name: name
        type: string
      - description: 试题状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ReadingPartListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取阅读篇列表
      tags:
      - Reading
  /config/reading-part/update:
    put:
      consumes:
      - application/json
      description: 更新阅读part
      parameters:
      - description: 阅读part内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.ReadingPartItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新阅读part
      tags:
      - Reading
  /config/reading/add:
    post:
      consumes:
      - application/json
      description: 新增阅读套题
      parameters:
      - description: 阅读套题内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.BasicReadingItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增阅读套题
      tags:
      - Reading
  /config/reading/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除阅读套题
      parameters:
      - description: 阅读套题ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除阅读套题
      tags:
      - Reading
  /config/reading/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取阅读详情
      parameters:
      - description: 阅读id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ReadingItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取阅读套题详情
      tags:
      - Reading
  /config/reading/list:
    get:
      consumes:
      - application/json
      description: 根据条件获取阅读列表，并返回分页结果
      parameters:
      - description: 阅读名称
        in: query
        name: name
        type: string
      - description: 阅读状态
        in: query
        name: status
        type: integer
      - description: 试题类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageLimit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.ReadingListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取阅读套题列表
      tags:
      - Reading
  /config/reading/update:
    put:
      consumes:
      - application/json
      description: 更新阅读套题
      parameters:
      - description: 阅读套题内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.BasicReadingItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 更新阅读套题
      tags:
      - Reading
  /config/testing/add:
    post:
      consumes:
      - application/json
      description: 新增测试套题
      parameters:
      - description: 测试套题内容
        in: body
        name: part
        required: true
        schema:
          $ref: '#/definitions/models.BasicTestingItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增测试套题
      tags:
      - Testing
  /config/testing/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除测试套题
      parameters:
      - description: 测试套题ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除测试套题
      tags:
      - Testing
  /config/testing/detail/{id}:
    get:
      consumes:
      - application/json
      description: 根据id获取测试详情
      parameters:
      - description: 测试id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/models.TestingItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                data:
                  type: object
              type: object
      summary: 获取测试套题详情
      tags:
      - Testing
...
swagger: "2.0"